Adiciona mais testes na camada da API
